Refactoring iam probe. Adding background as per #332
